+++ GUI/Search/BoxFrame/BoxFrame.lua
@@ -77,15 +77,15 @@ local function SearchAchievements(text, numAchievementIds, results, excludeExclu
 		achievement = addon.Data.Achievements[addon.Data.AchievementIds[i]];
 		local id, name, _, _, _, _, _, description, _, _, rewardText, _, _, _, _ = GetAchievementInfo(achievement.Id);
 		if id then
-			if (addon.SearchOptions.db.SearchIds and string.find(tostring(id), text, 1, true))
-			or (addon.SearchOptions.db.SearchNames and string.find(name:lower(), text, 1, true))
-			or (addon.SearchOptions.db.SearchDescriptions and string.find(description:lower(), text, 1, true))
-			or (addon.SearchOptions.db.SearchRewards and string.find(rewardText:lower(), text, 1, true)) then
+			if (addon.SearchOptions.db.profile.SearchIds and string.find(tostring(id), text, 1, true))
+			or (addon.SearchOptions.db.profile.SearchNames and string.find(name:lower(), text, 1, true))
+			or (addon.SearchOptions.db.profile.SearchDescriptions and string.find(description:lower(), text, 1, true))
+			or (addon.SearchOptions.db.profile.SearchRewards and string.find(rewardText:lower(), text, 1, true)) then
 				if not (excludeExcluded and achievement.Excluded) then
 					local value = 1;
 					if addon.Options.db.profile.SearchBox.OnlySearchFiltered then
 						local category;
-						if addon.Filters.db.MergeSmallCategories then
+						if addon.Filters.db.profile.MergeSmallCategories then
 							category = achievement:GetMergedCategory(); -- This way we get the parent category
 						else
 							category = achievement.Category;
@@ -166,7 +166,7 @@ local function GetSearchResults(text)
 
 	local numAchievementIds = #addon.Data.AchievementIds;
 	local excludeExcluded = addon.Options.db.profile.SearchBox.ExcludeExcluded;
-	local showPlaceholders = addon.Options.db.profile.ShowPlaceholdersFilter and addon.Filters.db.ShowPlaceholders;
+	local showPlaceholders = addon.Options.db.profile.ShowPlaceholdersFilter and addon.Filters.db.profile.ShowPlaceholders;
 
 	if string.match(text, "^#") then
 		results = SearchAchievementIds(text, numAchievementIds, results, excludeExcluded, showPlaceholders);
@@ -174,7 +174,7 @@ local function GetSearchResults(text)
 		results = SearchCriteria(text, numAchievementIds, results, excludeExcluded);
 	else
 		results = SearchAchievements(text, numAchievementIds, results, excludeExcluded);
-		if addon.SearchOptions.db.SearchCriteria then
+		if addon.SearchOptions.db.profile.SearchCriteria then
 			results = SearchCriteria(text, numAchievementIds, results, excludeExcluded);
 		else
 			table.sort(results, function(a, b)